refactor(EventData): move EventData def to event_data
