+++ src/event/mod.rs
@@ -3,13 +3,7 @@ use serde::{Deserialize, Serialize};
 pub mod event_data;
 pub mod sections;
 
-use self::event_data::{
-    delegated::{DelegatedInceptionEvent, DelegatedRotationEvent},
-    inception::InceptionEvent,
-    interaction::InteractionEvent,
-    receipt::EventReceipt,
-    rotation::RotationEvent,
-};
+use self::event_data::EventData;
 
 #[derive(Serialize, Deserialize)]
 pub struct Event {
@@ -20,16 +14,4 @@ pub struct Event {
 
     #[serde(flatten)]
     pub event_data: EventData,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(tag = "ilk")]
-#[serde(rename_all = "lowercase")]
-pub enum EventData {
-    Icp(InceptionEvent),
-    Rot(RotationEvent),
-    Ixn(InteractionEvent),
-    Dip(DelegatedInceptionEvent),
-    Drt(DelegatedRotationEvent),
-    Rct(EventReceipt),
 }
