+++ src/event_message/mod.rs
@@ -5,8 +5,6 @@ use serde::{Deserialize, Serialize};
 /// Event Message
 ///
 /// An EventMessage represents any signed message involved in the KERI protocol
-/// All types share mandatory prefix, sn, sig config and signatures fields, but differ
-/// in the event data they carry
 #[derive(Serialize, Deserialize)]
 pub struct EventMessage {
     #[serde(flatten)]
@@ -18,6 +16,6 @@ pub struct EventMessage {
     /// Appended Signatures
     ///
     /// TODO in the recommended JSON encoding, the signatures are appended to the json body.
-    /// how do you do that in serde?
+    #[serde(skip_serializing)]
     pub signatures: Vec<Prefix>,
 }
